+++ Tests/SentryTests/Integrations/ViewHierarchy/SentryViewHierarchyIntegrationTests.swift
@@ -111,6 +111,27 @@ class SentryViewHierarchyIntegrationTests: XCTestCase {
         XCTAssertEqual(newAttachmentList?.count, 0)
     }
 #endif // os(iOS) || targetEnvironment(macCatalyst)
+    
+    func test_noViewHierarchy_WhenDiscardedInCallback() {
+        let sut = fixture.getSut()
+
+        let expectation = expectation(description: "BeforeCaptureViewHierarchy must be called.")
+
+        let options = Options()
+        options.attachViewHierarchy = true
+        options.beforeCaptureViewHierarchy = { _ in
+            expectation.fulfill()
+            return false
+        }
+
+        sut.install(with: options)
+
+        let newAttachmentList = sut.processAttachments([], for: Event(error: NSError(domain: "", code: -1)))
+
+        wait(for: [expectation], timeout: 1.0)
+
+        XCTAssertEqual(newAttachmentList?.count, 0)
+    }
 
     func test_noViewHierarchy_keepAttachment() {
         let sut = fixture.getSut()